Fix for mangage user error msg issue

+++ admin/manage-users.php
@@ -23,28 +23,6 @@ $users=mysqli_query($connection,$query);
             </p>
             
             </div>
-            <?php
-        // elseif(isset($_SESSION['edit-user'])): 
-        // ?> 
-        //     <div class="alert__message error container">
-        //     <p>
-        //         <?= $_SESSION['edit-user'];
-        //         unset($_SESSION['edit-user']);
-        //         ?>
-        //     </p>
-            
-        //     </div>
-        <!-- <?php
-        // elseif(isset($_SESSION['edit-user-success'])): 
-        ?> 
-            <div class="alert__message success container">
-            <p>
-                <?= $_SESSION['edit-user-success'];
-                unset($_SESSION['edit-user-success']);
-                ?>
-            </p>
-            
-            </div> -->
         <?php
             elseif(isset($_SESSION['delete-user'])): 
         ?> 
@@ -78,12 +56,12 @@ $users=mysqli_query($connection,$query);
                 </p>
                 <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
             </div>
-        <?php elseif(isset($_SESSION['edit-user'])): ?>
+        <?php elseif(isset($_SESSION['edit-user-fail'])): ?>
 
             <div class="alert alert-danger alert-dismissible fade show alert-settings" role="alert">
                 <p>
-                    <?=$_SESSION['edit-user'];
-                    unset($_SESSION['edit-user']); 
+                    <?=$_SESSION['edit-user-fail'];
+                    unset($_SESSION['edit-user-fail']); 
                     ?>
                 </p>
                 <button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>
@@ -113,10 +91,8 @@ $users=mysqli_query($connection,$query);
                             <td><?= $user["username"]  ?> </td>
                             <td class="text-center"><?= $user["is_admin"] ? 'Yes' : 'No'  ?></td>
                             <td class="text-center"><a href="<?= ROOT_URL ?>admin/edit-user.php?id=<?= $user['id'] ?>" class="btn btn-primary"><i class="fa fa-edit"></i>&nbsp; Edit</a></td>
-                            <td class="text-center"><a href="<?= ROOT_URL ?>admin/delete-users.php?id=<?= $user['id'] ?>" class="btn btn-danger"><i class="fa fa-trash"></i>&nbsp;Delete</a></td>
-                            
-
-
+                            <td class="text-center"><a href="<?= ROOT_URL ?>admin/delete-users.php?id=<?= $user['id'] ?>" class="btn btn-danger"><i class="fa fa-trash" onclick="return confirm('Are you sure?');"></i>&nbsp;Delete</a></td>
+                            <td class="text-center"><a href="<?= ROOT_URL ?>admin/delete-post.php?id=<?= $post['id'] ?>" class="btn btn-danger" onclick="return confirm('Are you sure?');"><i class="fa fa-trash"></i>&nbsp;Delete</a></td>
                         </tr>
                         <?php endwhile ?>
